+++ src/main/java/com/umc/i/src/member/MemberDao.java
@@ -1,4 +1,4 @@
-package com.umc.i.src.member.join.auth;
+package com.umc.i.src.member;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
@@ -9,11 +9,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Repository;
 
-import com.umc.i.src.member.join.model.PostJoinReq;
+import com.umc.i.src.member.model.post.PostJoinAuthReq;
 
 @Repository
 
-public class AuthDao {
+public class MemberDao {
     private JdbcTemplate jdbcTemplate;
 
     @Autowired
@@ -26,13 +26,13 @@ public class AuthDao {
     //이메일 중복 확인
 
     //인증 코드 발송(인증테이블에 저장)
-    public String createAuth(PostJoinReq postJoinReq, String key) {
+    public String createAuth(PostJoinAuthReq postJoinAuthReq, String key) {
         LocalDateTime time = LocalDateTime.now();
         DateTimeFormatter timeFormatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
         String currentTime = time.format(timeFormatter);
         
         String createAuthQuery = "insert into Member_auth (ma_key, ma_type, ma_generate_time, ma_expired) values (?, ?, ?, ?)";
-        Object[] createAuthParams = new Object[]{key, postJoinReq.getType(), currentTime, false}; // 동적 쿼리의 ?부분에 주입될 값
+        Object[] createAuthParams = new Object[]{key, postJoinAuthReq.getType(), currentTime, false}; // 동적 쿼리의 ?부분에 주입될 값
         this.jdbcTemplate.update(createAuthQuery, createAuthParams);    //인증 정보 생성
 
         String lastInserIdQuery = "select last_insert_id()"; // 가장 마지막에 삽입된(생성된) id값은 가져온다.
