+++ src/main/java/com/umc/i/src/member/MemberService.java
@@ -1,4 +1,4 @@
-package com.umc.i.src.member.join.auth;
+package com.umc.i.src.member;
 
 import java.io.BufferedReader;
 import java.io.DataOutputStream;
@@ -13,19 +13,28 @@ import java.util.Random;
 
 import javax.crypto.Mac;
 import javax.crypto.spec.SecretKeySpec;
+import javax.mail.MessagingException;
+import javax.mail.internet.MimeMessage;
 
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
+import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.stereotype.Service;
+import org.thymeleaf.context.Context;
+import org.thymeleaf.spring5.SpringTemplateEngine;
 
 import lombok.RequiredArgsConstructor;
 
 @Service
 @RequiredArgsConstructor
-public class PhoneAuthService {
+public class MemberService {
+    //의존성 주입 -> 필요한 객체 가져오기
+    private final JavaMailSender eMailSender;
+    // 타임리프를 사용하기 위한 객체
+    private final SpringTemplateEngine templateEngine;
     private String authCode; //인증코드
 
-    //인증코드 생성
+    // 랜덤 인증코드 생성
     public void createCode() {
         Random random = new Random();
         int checkNum = random.nextInt(899999) + 100000; // 인증번호 범위: 100000 ~ 999999
@@ -33,7 +42,40 @@ public class PhoneAuthService {
         authCode = Integer.toString(checkNum);
     }
 
-	@SuppressWarnings("unchecked")
+    // 메일 양식 작성
+    public MimeMessage createEmailForm(String email) throws MessagingException, UnsupportedEncodingException {
+        createCode();   // 인증코드 생성
+        String setFrom = "amanda010926@gmail.com";  // 보내는 사람 이메일
+        String toEmail = email; //받는 사람
+        String title = "아이 - 아름답게 이별하는 법 본인 인증 코드";    // 메일 제목
+
+        MimeMessage message = eMailSender.createMimeMessage();
+        message.addRecipients(MimeMessage.RecipientType.TO, toEmail);     //보낼 이메일 설정
+        message.setSubject(title);      // 제목 설정   
+        message.setFrom(setFrom);       // 보내는 이메일
+        message.setText(setContext(authCode), "utf-8", "html");
+
+        return message;
+    }
+
+    // 메일 전송
+    public String sendEmail(String toEmail) throws MessagingException, UnsupportedEncodingException {
+        // 메일 전송에 필요한 정보 설정
+        MimeMessage emailForm = createEmailForm(toEmail);
+        // 실제 메일 전송
+        eMailSender.send(emailForm);
+
+        return authCode;
+    }
+
+    // 타임리프를 이용한 context 설정
+    public String setContext(String code) {
+        Context context = new Context();
+        context.setVariable("code", code);
+        return templateEngine.process("signup", context);     // signup.html
+    }
+
+    @SuppressWarnings("unchecked")
 	public String send_msg(String tel) {
         String hostNameUrl = "https://sens.apigw.ntruss.com";     		// 호스트 URL
         String requestUrl= "/sms/v2/services/";                   		// 요청 URL
@@ -68,7 +110,6 @@ public class PhoneAuthService {
 	    System.out.println(body);
 	    
         try {
-
             URL url = new URL(apiUrl);
 
             HttpURLConnection con = (HttpURLConnection)url.openConnection();
@@ -142,4 +183,5 @@ public class PhoneAuthService {
 
 	  return encodeBase64String;
 	}
+    
 }
