+++ src/main/java/com/umc/i/src/member/model/post/PostJoinAuthReq.java
@@ -1,4 +1,4 @@
-package com.umc.i.src.member.join.model;
+package com.umc.i.src.member.model.post;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -10,7 +10,7 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class PostJoinReq {
+public class PostJoinAuthReq {
     private int type;   // 1: mail, 2: phone
     private String auth;    // 메일주소, 핸드폰 번호
 }
